+++ YammyYammyShop_App2023/Model/Position.swift
@@ -11,9 +11,9 @@ import FirebaseFirestore
 struct Position: Identifiable {
     var id: String
     var product: Product
-    var count: Double
+    var count: Int
     
-    var cost: Double {
+    var cost: Int {
         return product.price * self.count
     }
     
@@ -28,7 +28,7 @@ struct Position: Identifiable {
         return repres
     }
     
-    internal init(id: String, product: Product, count: Double) {
+    internal init(id: String, product: Product, count: Int) {
         self.id = id
         self.product = product
         self.count = count
@@ -39,14 +39,14 @@ struct Position: Identifiable {
         
         guard let id = data["id"] as? String else { return nil }
         guard let title = data["title"] as? String else { return nil }
-        guard let price = data["price"] as? Double else { return nil }
+        guard let price = data["price"] as? Int else { return nil }
         
         let product: Product = Product(id: "",
                                        title: title,
                                        imageUrl: "",
                                        price: price,
                                        description: "")
-        guard let count = data["count"] as? Double else { return nil }
+        guard let count = data["count"] as? Int else { return nil }
         
         self.id = id
         self.product = product
